+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Changelog
 
-# 
+# 0.20.0rc0
 ## No Breaking Changes
 
 ## Features Added
-  - Release notes generation.
+  - Support has been added for topic inputs to action invocations, via `InvocationInput.topics`. Action writers can access the topics via `ActionInput.topics`.
+  - Added the `requires_downloaded_inputs` optional flag to `Action::create` and `Action::update`. It controls whether an action invocation's inputs will be available in its working directory before business logic runs. This is true by default.
+  - Added getters to `Action` for `description`, `short_description`, `tags`, `metadata`, `published` and `requires_downloaded_inputs`.
+  - `Dataset::get_topics` now has optional arguments to `include` or `exclude` topics by name, similar to `File::get_topics`.
+  - Added `FileDownloader` to simplify the task of downloading multiple files, for instance from search results.
+  - Added `CanonicalDataType.Timestamp` to support identifying `MessagePath`s that should be interpreted as time elapsed since the Unix epoch.
 
 ## Bugs Fixed
 
@@ -31,7 +36,7 @@
   - Added `TriggerEvaluationCause::FileIngest` to allow triggers to respond to when a file is marked as `IngestionStatus::Ingested`.
 
 ## Bugs Fixed
-  - Added dynamic import guard around `roboto.version` in `requester.py` to fix in-IDE tests from Roboto's development environment (before `version.py` is dynamically generated)
+  - Added dynamic import guard around `roboto.version` in `requester.py` to fix in-IDE tests from Roboto's development environment (before `version.py` is dynamically generated).
   - Don't reject API requests that contain extra fields. This enables backwards-compatibility with outdated SDK builds and forwards-compatibility for adding new fields to the SDK independent of our server release cycle.
 
 # 0.17.0